% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{data_summary}
\alias{data_summary}
\title{Generate a summary report for numeric variables}
\usage{
data_summary(data)
}
\arguments{
\item{data}{This argument represents the data frame containing the dataset to be summarized, which must contain at least one numeric column; or, the function will throw an error.}
}
\value{
The function returns a list containing: the mean, median, standard deviation and interquartile range for the numerical variables
}
\description{
Given a data frame, 'data_summary()' summarizes all numerical values in the data set and returns a list.
The function summarizes the mean, median, standard deviation and interquartile range for all columns containing numeric data.
}
\details{
The 'data_summary' function provides a summary report for a dataset generating summary statistics (mean, median, standard deviation and interquartile range) for numerical variables.
}
\examples{
# example dataset
df <- data.frame(
category = c("A", "B", "C"),
value1 = c(10, 17, 28),
value2 = c(15, 12, 20))
data_summary(df)
}
